refactor: share the CGEvent tap lifecycle in EventTap
