+++ scripts/line-limit-baseline
@@ -10,6 +10,5 @@
 # lines  path
 437   App/Tests/plonkTests/ConfigTests.swift
 444   App/Tests/plonkTests/RouterTests.swift
-312   App/Sources/plonk/GrabMove.swift
 308   App/Sources/plonk/WorkspaceLauncher.swift
 306   mcp/src/api.ts
